Debug a retain cycle; moved ivars to implementation; cleared all warnings generated using my larger set of warnings

+++ Core/Source/NSString+HTML.m
@@ -54,7 +54,7 @@ static NSDictionary *entityReverseLookup = nil;
 {
 	const char *s = [self UTF8String];
 	
-	for (int i=0;i<strlen(s);i++)
+	for (size_t i=0;i<strlen(s);i++)
 	{
 		if (s[i]<'0' || s[i]>'9')
 		{
@@ -70,7 +70,7 @@ static NSDictionary *entityReverseLookup = nil;
 	float result = 1;
 	sscanf([self UTF8String], "%f", &result);
 	
-	return result/100.0;
+	return result/100.0f;
 }
 
 - (NSString *)stringByNormalizingWhitespace
@@ -394,7 +394,7 @@ static NSDictionary *entityReverseLookup = nil;
 	
 	NSMutableString *tmpString = [NSMutableString string];
 	
-	for (int i = 0; i<[self length]; i++)
+	for (NSUInteger i = 0; i<[self length]; i++)
 	{
 		unichar oneChar = [self characterAtIndex:i];
 		
@@ -730,7 +730,7 @@ static NSDictionary *entityReverseLookup = nil;
 	
 	NSMutableString *tmpString = [NSMutableString string];
 	
-	for (int i = 0; i<[self length]; i++)
+	for (NSUInteger i = 0; i<[self length]; i++)
 	{
 		unichar oneChar = [self characterAtIndex:i];
 		
@@ -777,7 +777,7 @@ static NSDictionary *entityReverseLookup = nil;
 	}
 	else if ([self hasSuffix:@"%"])
 	{
-		return value * textSize / 100.0;
+		return value * textSize / 100.0f;
 	}
 	
 	// everything else interpret as pixels
